Fix the weird orb bug
